Add Loading Spinner & Timeout Widget
This commit adds a loading spinner to each canteen stack which gets pushed down with new day entries until all available days are displayed.
If no menu could be fetched an error message will be shown instead to the user, this error message should in future inform the user to reload the
content when an internet connection is available.